+++ source/lib/rocprofiler-sdk/context.cpp
@@ -79,6 +79,9 @@ rocprofiler_stop_context(rocprofiler_context_id_t context_id)
        !rocprofiler::context::get_registered_context(context_id))
         return ROCPROFILER_STATUS_ERROR_CONTEXT_NOT_FOUND;
 
+    // if finalized, context is already stopped
+    if(rocprofiler::registration::get_fini_status() > 0) return ROCPROFILER_STATUS_SUCCESS;
+
     return rocprofiler::context::stop_context(context_id);
 }