+++ CHANGELOG.md
@@ -147,7 +147,9 @@ Full documentation for ROCprofiler-SDK is available at [rocm.docs.amd.com/projec
 ### Added
 
 - Added support for select() operation in counter expression.
-- Added reduce operation for counter expression wrt dimension. 
+- Added reduce operation for counter expression wrt dimension.
+- `--collection-period` feature added in rocprofv3, to enable filtering using time.
+- `--collection-period-unit` feature added in rocprofv3, to allow the user to control time units used in collection period option.
 
 ### Changed